add queryset tag context builder
